backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
openpyxl==3.1.5
django-filter==24.3
python-dateutil==2.9.0